Update On-prem infra model
